+++ src/swap/central/sideshift.ts
@@ -1,7 +1,17 @@
-import { asBoolean, asEither, asObject, asOptional, asString } from 'cleaners'
+import {
+  asArray,
+  asBoolean,
+  asEither,
+  asMaybe,
+  asObject,
+  asOptional,
+  asString,
+  asUnknown
+} from 'cleaners'
 import {
   EdgeCorePluginOptions,
   EdgeFetchFunction,
+  EdgeLog,
   EdgeMemo,
   EdgeSpendInfo,
   EdgeSwapInfo,
@@ -16,12 +26,12 @@ import {
 } from 'edge-core-js/types'
 
 import {
-  checkInvalidCodes,
+  ChainCodeTickerMap,
   checkWhitelistedMainnetCodes,
+  CurrencyPluginIdSwapChainCodeMap,
   ensureInFuture,
-  getCodesWithTranscription,
+  getChainAndTokenCodes,
   getMaxSwappable,
-  InvalidCurrencyCodes,
   makeSwapPluginQuote,
   SwapOrder
 } from '../../util/swapHelpers'
@@ -29,79 +39,70 @@ import { convertRequest, getAddress, memoType } from '../../util/utils'
 import { EdgeSwapRequestPlugin } from '../types'
 
 // See https://help.sideshift.ai/en/articles/4559664-which-coins-and-tokens-are-listed for list of supported currencies
-const MAINNET_CODE_TRANSCRIPTION = {
+export const MAINNET_CODE_TRANSCRIPTION: CurrencyPluginIdSwapChainCodeMap = {
   algorand: 'algorand',
   arbitrum: 'arbitrum',
   avalanche: 'avax',
-  // axelar:
+  axelar: null,
   base: 'base',
-  // binance:
+  binance: null,
   binancesmartchain: 'bsc',
   bitcoin: 'bitcoin',
   bitcoincash: 'bitcoincash',
-  // bitcoingold:
+  bitcoingold: null,
   bitcoinsv: 'bsv',
+  bobevm: null,
   cardano: 'cardano',
-  // celo:
-  // coreum:
+  celo: null,
+  coreum: null,
   cosmoshub: 'cosmos',
   dash: 'dash',
-  // digibyte:
+  digibyte: null,
   dogecoin: 'doge',
-  // eboost:
-  // eos:
+  eboost: null,
+  eos: null,
   ethereum: 'ethereum',
   ethereumclassic: 'etc',
-  // ethereumpow:
+  ethereumpow: null,
   fantom: 'fantom',
-  // feathercoin:
-  // filecoin:
-  // filecoinfevm:
-  // fio:
+  feathercoin: null,
+  filecoin: null,
+  filecoinfevm: null,
+  fio: null,
   groestlcoin: 'grs',
-  // hedera:
-  // liberland:
+  hedera: null,
+  liberland: null,
   litecoin: 'litecoin',
   monero: 'monero',
   optimism: 'optimism',
-  // osmosis:
-  // piratechain:
+  osmosis: null,
+  piratechain: null,
   polkadot: 'polkadot',
   polygon: 'polygon',
-  // pulsechain:
-  // qtum:
-  // ravencoin:
+  pulsechain: null,
+  qtum: null,
+  ravencoin: null,
   ripple: 'ripple',
-  // rsk:
-  // smartcash:
+  rsk: 'rootstock',
+  smartcash: null,
   solana: 'solana',
   stellar: 'stellar',
-  // telos:
+  telos: null,
   tezos: 'tezos',
-  // thorchainrune:
+  thorchainrune: null,
   ton: 'ton',
   tron: 'tron',
-  // ufo:
-  // vertcoin:
-  // wax:
+  ufo: null,
+  vertcoin: null,
+  wax: null,
   zcash: 'shielded',
-  // zcoin:
+  zcoin: null,
   zksync: 'zksyncera'
-}
-
-const INVALID_CURRENCY_CODES: InvalidCurrencyCodes = {
-  from: {
-    optimism: ['VELO'],
-    polygon: ['USDC', 'USDC.e']
-  },
-  to: {
-    polygon: ['USDC', 'USDC.e']
-  }
 }
 
 const SIDESHIFT_BASE_URL = 'https://sideshift.ai/api/v2'
 const pluginId = 'sideshift'
-const swapInfo: EdgeSwapInfo = {
+export const swapInfo: EdgeSwapInfo = {
   pluginId,
   isDex: false,
   displayName: 'SideShift.ai',
@@ -114,12 +115,12 @@ async function checkQuoteError(
   request: EdgeSwapRequestPlugin,
   quoteErrorMessage: string
 ): Promise<void> {
-  const { fromCurrencyCode, fromWallet } = request
+  const { fromWallet } = request
 
   if (quoteErrorMessage === 'Amount too low') {
     const nativeMin = await fromWallet.denominationToNative(
       rate.min,
-      fromCurrencyCode
+      request.fromCurrencyCode
     )
     throw new SwapBelowLimitError(swapInfo, nativeMin)
   }
@@ -127,7 +128,7 @@ async function checkQuoteError(
   if (quoteErrorMessage === 'Amount too high') {
     const nativeMax = await fromWallet.denominationToNative(
       rate.max,
-      fromCurrencyCode
+      request.fromCurrencyCode
     )
     throw new SwapAboveLimitError(swapInfo, nativeMax)
   }
@@ -200,7 +201,12 @@ const fetchSwapQuoteInner = async (
     toCurrencyCode,
     fromMainnetCode,
     toMainnetCode
-  } = getCodesWithTranscription(request, MAINNET_CODE_TRANSCRIPTION)
+  } = await getChainAndTokenCodes(
+    request,
+    swapInfo,
+    chainCodeTickerMap,
+    MAINNET_CODE_TRANSCRIPTION
+  )
 
   const rate = asRate(
     await api.get<typeof asRate>(
@@ -332,10 +338,75 @@ const fetchSwapQuoteInner = async (
   }
 }
 
-const createFetchSwapQuote = (api: SideshiftApi, affiliateId: string) =>
+// Provider data
+let chainCodeTickerMap: ChainCodeTickerMap = new Map()
+let lastUpdated = 0
+const EXPIRATION = 1000 * 60 * 60 // 1 hour
+
+async function fetchSupportedAssets(
+  api: CreateSideshiftApiResponse,
+  log: EdgeLog
+): Promise<void> {
+  if (lastUpdated > Date.now() - EXPIRATION) return
+
+  try {
+    const json = await api.get('/coins')
+    const assets = asSideshiftAssets(json)
+
+    const chaincodeArray = Object.values(MAINNET_CODE_TRANSCRIPTION)
+    const out: ChainCodeTickerMap = new Map()
+    for (const asset of assets) {
+      const mainnetObj = asMaybe(asMainnetAsset)(asset)
+      if (mainnetObj != null) {
+        for (const network of mainnetObj.networks) {
+          if (chaincodeArray.includes(network)) {
+            const tokenCodes = out.get(network) ?? []
+            tokenCodes.push({
+              tokenCode: mainnetObj.coin,
+              contractAddress: null
+            })
+            out.set(network, tokenCodes)
+          }
+        }
+      }
+
+      const tokenObj = asMaybe(asTokenAsset)(asset)
+      if (tokenObj != null) {
+        for (const network of tokenObj.networks) {
+          if (chaincodeArray.includes(network)) {
+            const tokenCodes = out.get(network) ?? []
+            const networkObj = Object.keys(tokenObj.tokenDetails).find(
+              networkName => networkName === network
+            )
+            if (networkObj == null) continue
+            tokenCodes.push({
+              tokenCode: tokenObj.coin,
+              contractAddress: tokenObj.tokenDetails[networkObj].contractAddress
+            })
+            out.set(network, tokenCodes)
+          }
+        }
+      }
+    }
+
+    chainCodeTickerMap = out
+    lastUpdated = Date.now()
+  } catch (e) {
+    log.warn('SideShift: Error updating supported assets', e)
+  }
+}
+
+const createFetchSwapQuote = (
+  api: SideshiftApi,
+  affiliateId: string,
+  log: EdgeLog
+) =>
   async function fetchSwapQuote(req: EdgeSwapRequest): Promise<EdgeSwapQuote> {
     const request = convertRequest(req)
-    checkInvalidCodes(INVALID_CURRENCY_CODES, request, swapInfo)
+
+    // Fetch and persist chaincode/tokencode maps from provider
+    await fetchSupportedAssets(api, log)
+
     checkWhitelistedMainnetCodes(MAINNET_CODE_TRANSCRIPTION, request, swapInfo)
 
     const newRequest = await getMaxSwappable(
@@ -353,7 +424,11 @@ export function makeSideshiftPlugin(
 ): EdgeSwapPlugin {
   const { io, initOptions } = opts
   const api = createSideshiftApi(SIDESHIFT_BASE_URL, io.fetchCors ?? io.fetch)
-  const fetchSwapQuote = createFetchSwapQuote(api, initOptions.affiliateId)
+  const fetchSwapQuote = createFetchSwapQuote(
+    api,
+    initOptions.affiliateId,
+    opts.log
+  )
 
   return {
     swapInfo,
@@ -426,3 +501,19 @@ const asOrder = asEither(
   }),
   asError
 )
+
+const asMainnetAsset = asObject({
+  networks: asArray(asString),
+  coin: asString,
+  mainnet: asString
+})
+const asTokenAsset = asObject({
+  networks: asArray(asString),
+  coin: asString,
+  tokenDetails: asObject(
+    asObject({
+      contractAddress: asString
+    })
+  )
+})
+const asSideshiftAssets = asArray(asUnknown)